requirements.txt
streamlit
python-dotenv
pandas
nadeo-api